Fix imports etc. after refactoring
