+++ src/deflex/postprocessing/analyses.py
@@ -252,90 +252,6 @@ class Cycles:
                 print("")
 
 
-def get_all_nodes_from_results(results):
-    keys = sorted(list(results["main"].keys()))
-    unique_nodes = []
-    for nodes in keys:
-        unique_nodes.append(nodes[0])
-        if nodes[1] is not None:
-            unique_nodes.append(nodes[1])
-    return set(unique_nodes)
-
-
-def nodes2table(results, no_sums=False):
-    """
-    Get a table with all nodes as a MultiIndex with the sum of their in an out
-    flows.
-
-    The index contains the following levels: class, category, tag, subtag,
-    region
-
-    The sums can be found in the columns "in" and "out".
-
-    Parameters
-    ----------
-    results : dict
-        Deflex results dictionary.
-    no_sums : bool
-        Set to False to get an empty DataFrame with no sums (default: True)
-
-    Returns
-    -------
-    Table with all nodes and sums : pandas.DataFrame
-
-    Examples
-    --------
-    >>> from deflex import tools
-    >>> fn = tools.fetch_test_files("de03_fictive.dflx")
-    >>> my_results = tools.files.restore_results(fn)
-    >>> all_nodes = nodes2table(my_results)
-    >>> len(all_nodes)
-    226
-    >>> all_nodes.to_csv("your/path/file.csv")  # doctest: +SKIP
-
-
-    """
-    unique_nodes = get_all_nodes_from_results(results)
-    nodes = []
-    for node in unique_nodes:
-        dc = {}
-        solph_class = type(node)
-        label = node.label
-        dc["class"] = (
-            str(solph_class).rsplit(".", maxsplit=1)[-1].replace("'>", "")
-        )
-        dc["cat"] = label.cat
-        dc["tag"] = label.tag
-        dc["subtag"] = label.subtag
-        dc["region"] = label.region
-        if no_sums is False:
-            from_node = sum(
-                [
-                    v["sequences"]["flow"]
-                    for k, v in results["Main"].items()
-                    if k[0].label == label and k[1] is not None
-                ]
-            )
-            if isinstance(from_node, pd.Series):
-                from_node = from_node.sum()
-            to_node = sum(
-                [
-                    v["sequences"]["flow"]
-                    for k, v in results["Main"].items()
-                    if getattr(k[1], "label", "") == label
-                ]
-            )
-            if isinstance(to_node, pd.Series):
-                to_node = to_node.sum()
-            dc["out"] = from_node
-            dc["in"] = to_node
-        nodes.append(dc)
-    df = pd.DataFrame(nodes)
-    df.sort_values(by=list(df.columns), inplace=True)
-    df.reset_index(drop=True, inplace=True)
-    return df.set_index(["class", "cat", "tag", "subtag", "region"], drop=True)
-
-
 def get_resource_parameters(results, bus):
     """
     Get the parameters of a commodity bus.
@@ -443,12 +359,6 @@ def fetch_converter_parameters(results, transformer):
                 "conversion_factors_{}".format(label2str(outflow[1].label))
             ]
 
-    # # Fetch efficiency of heat plants as reference efficiency for chp.
-    # for i, row in df.loc[df.category == "chp plant"].iterrows():
-    #     df.loc[i, "efficiency, hp_ref"] = df.loc[
-    #         (df.fuel == row.fuel) & (df.category == "heat plant"),
-    #         "efficiency, heat",
-    #     ].mean()
     return df.sort_index(axis=1)
 
 
@@ -532,7 +442,6 @@ def fetch_electricity_flows(results):
             and k[1].label.cat == "electricity"
         }
     )
-    # flow_status = flows.div(flows).fillna(0)
 
 
 def calculate_key_values(results):
@@ -559,7 +468,6 @@ def calculate_key_values(results):
     flow_status = flows.div(flows).fillna(0)
 
     converter_parameters = calculate_marginal_costs(converter_parameters)
-    # em_max = flow_status.mul(converter_parameters["emissions"]).max(1)
 
     kv = pd.DataFrame()
