+++ tests/test_analyses.py
@@ -34,7 +34,7 @@ class TestAnalysis:
         cls.results = restore_results(fetch_test_files("de02_no-heat.dflx"))
 
     def test_all_nodes_from_results(self):
-        nodes = postprocessing.analyses.get_all_nodes_from_results(
+        nodes = postprocessing.basic.get_all_nodes_from_results(
             self.results
         )
         assert len(nodes) == 141
@@ -48,7 +48,7 @@ class TestAnalysis:
 
     def test_graph_and_nodes(self):
         graph = postprocessing.DeflexGraph(self.results)
-        nodes = postprocessing.analyses.get_all_nodes_from_results(
+        nodes = postprocessing.basic.get_all_nodes_from_results(
             self.results
         )
         n = 0
@@ -74,7 +74,7 @@ class TestAnalysis:
         )
 
     def test_resource_parameter(self):
-        nodes = postprocessing.analyses.get_all_nodes_from_results(
+        nodes = postprocessing.basic.get_all_nodes_from_results(
             self.results
         )
         buses = [
@@ -97,7 +97,7 @@ class TestAnalysis:
         assert round(emission / number, 4) == 0.1804
 
     def test_converter_parameters(self):
-        nodes = postprocessing.analyses.get_all_nodes_from_results(
+        nodes = postprocessing.basic.get_all_nodes_from_results(
             self.results
         )
         transformer = [
@@ -121,7 +121,7 @@ class TestAnalysis:
         postprocessing.fetch_parameter_of_commodity_sources(self.results)
 
     def test_marginal_costs(self):
-        nodes = postprocessing.analyses.get_all_nodes_from_results(
+        nodes = postprocessing.basic.get_all_nodes_from_results(
             self.results
         )
         transformer = [
